Refactoring
Move compute_rij_wij into aux_cosym and provide for unit testing
